+++ src/constants.js
@@ -1,2 +1,5 @@
 // Database
 export const DB_FILE = 'app-data.db'
+
+// Media
+export const MEDIA_DIR = 'Cronos'